refactor: Update buildHTML function to accept an array of Review objects
The buildHTML function in utils.ts has been updated to accept an array of Review objects instead of an array of numbers. This change allows for more flexibility in the data that can be plotted in the HTML output. The function now takes an array of Review objects, each containing a rating and a concordance value. The ratings and concordance values are used to generate separate lines on the plot.

This change improves the readability and maintainability of the code by providing a more descriptive and accurate representation of the data being plotted.

Refactor the buildHTML function to accept an array of Review objects instead of an array of numbers.